initial port of 1/2 of twp.instancer to new stuff
This is the instance building part:

* create a new instance
* establish the config
* create the basic bags, recipes and users
* fill up with default tiddlers

The difference in this version:

* tiddlers are accessed via twp.pkgstore
* the store_contents concept is gone

There's some ambiguity in that instance_pkgstores is defined
in config while the other instance information is *.instance.

This mirrors how things were in instancer.

I _think_ it is probably correct as that information is
used by the update command and update can happen long after
instance creation _and_ instance_pkgstores is a list which
can contain several packages, include none that have anything
to do with the _definition_ of the current package's instance.